refactor(tests): move InteractionDefinition in own module
The `__init__` was getting a bit too cluttered, so I have refactored the
entire `InteractionDefinition` class into its own module. I have also
refacted the nested classes to stand alone now.

A minor refactor was also made to avoid assigning custom attributes to
the `Flask` app, and some changes were made to the provider
initialisation to avoid passing a redundant `self.app` in a method call.
